pyproject.toml
[project]
name = "infokanal-backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Rami Borni", email = "rami@scribfy.app" }
]
dependencies = [
    "Django==4.1.10",
    "pillow>=10.3.0",
    "PySocks==1.7.1",
    "aiohttp==3.8.4",
    "aiosignal==1.3.1",
    "anyio==3.7.1",
    "appdirs==1.4.4",
    "asgiref==3.7.2",
    "async-timeout==4.0.2",
    "attrs==23.1.0",
    "beautifulsoup4>=4.12.3",
    "bs4==0.0.1",
    "certifi==2023.5.7",
    "charset-normalizer==3.2.0",
    "click==8.1.5",
    "colorama==0.4.6",
    "cssselect==1.2.0",
    "django-cors-headers==4.2.0",
    "django-cron==0.6.0",
    "django-rest-framework==0.1.0",
    "djangorestframework==3.14.0",
    "djongo==1.3.6",
    "dnspython==2.4.0",
    "exceptiongroup==1.1.2",
    "fake-useragent==1.1.3",
    "feedgen==0.9.0",
    "feedparser==6.0.10",
    "frozenlist==1.4.0",
    "h11==0.14.0",
    "httpcore==0.17.3",
    "idna==3.4",
    "importlib-metadata==6.8.0",
    "joblib==1.3.1",
    "lxml==4.9.3",
    "multidict==6.0.4",
    "nltk==3.8.1",
    "numpy==1.25.1",
    "outcome==1.2.0",
    "pandas==2.0.3",
    "parse==1.19.1",
    "pyee==8.2.2",
    "pymongo==3.13.0",
    "pyppeteer==1.0.2",
    "pyquery==2.0.0",
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "regex==2023.6.3",
    "requests==2.31.0",
    "requests-html==0.10.0",
    "selenium==4.9.0",
    "sgmllib3k==1.0.0",
    "six==1.16.0",
    "sniffio==1.3.0",
    "sortedcontainers==2.4.0",
    "soupsieve==2.4.1",
    "sqlparse==0.2.4",
    "tqdm==4.65.0",
    "trio==0.22.2",
    "trio-websocket==0.10.3",
    "typing_extensions==4.7.1",
    "tzdata==2023.3",
    "urllib3==1.26.16",
    "w3lib==2.1.1",
    "websockets==10.4",
    "wsproto==1.2.0",
    "yarl==1.9.2",
    "zipp==3.16.2",
    "python-dotenv>=1.0.1",
    "langchain==0.1.16",
    "langchain-core==0.1.45",
    "langchain-community==0.0.32",
    "langchain-openai==0.1.1",
    "pydantic<2.0.0",
    "pytest-playwright>=0.4.4",
    "langchain-google-genai>=1.0.3",
    # Example: Check if there's a version of playwright that works with pyee==8.2.2
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/infokanal_backend"]
